+++ docs/app/(content)/components/[slug]/getApiReferenceData.ts
