+++ docs/src/components/content/MDXComponents.tsx
@@ -1,5 +1,6 @@
-import { ComponentLinkHeader, InstallationInstructions } from 'docs-base/src/modules/components';
-import { Callout } from 'docs-base/src/modules/common/Callout';
+import { ComponentLinkHeader } from './ComponentLinkHeader';
+import { InstallationInstructions } from './InstallationInstructions';
+import { Callout } from './Callout';
 import { PackageManagerSnippet, Npm, Pnpm, Yarn } from './PackageManagerSnippet';
 
 export const components = {
