+++ docs/src/components/content/PackageManagerSnippet.module.css
