+++ docs/src/components/content/PackageManagerSnippet.tsx
