--- docs/src/modules/components/index.ts
@@ -1,4 +1,0 @@
-export * from './ComponentLinkHeader';
-export * from './Demo';
-export * from './Description';
-export * from './InstallationInstructions';
